PowerShell profile settings
This function and alias help me use the command "touch .ignore" in PowerShell for a simple way to create a file in windows that starts with a period. I could use a period at the head and the tail of the name with ni or new-item but then I would need to add the flag to specify the item as a file or, without the flag, PS would need me to answer a query as to the Type: file.